refactor: extract useGTDFrames hook
Moves gtdFrames (frame definitions), showFramesModal, framesModalTab,
editingFrame, smartScheduleResults/Loading/Error/Accepted,
showRescheduleModal, rescheduleResults/Loading/Error/Accepted,
frameContextMenu, quickAddFrameModal, frameAdjustModal, frameAdjustTimeField,
and frameScheduleModal state into a dedicated useGTDFrames hook.

gtdFrames remains accessible in App.jsx scope so cloudSyncUpload can
reference it without any additional wiring.
